requirements.txt
flask
flask_smorest
python-dotenv
pymongo
openai